Fix new path
